+++ src/Agent/NewRelic/Agent/Core/AgentManager.cs
