+++ src/Agent/NewRelic/Agent/Core/AgentShim.cs
