+++ src/Agent/NewRelic/Agent/Core/DataTransport/DataTransportResponseStatus.cs
@@ -1,7 +1,7 @@
 // Copyright 2020 New Relic, Inc. All rights reserved.
 // SPDX-License-Identifier: Apache-2.0
 
-namespace NewRelic.Agent.Core
+namespace NewRelic.Agent.Core.DataTransport
 {
     public enum DataTransportResponseStatus
     {
