+++ src/Agent/NewRelic/Agent/Core/DisabledAgentManager.cs
