+++ src/Agent/NewRelic/Agent/Core/Environment.cs
