+++ src/Agent/NewRelic/Agent/Core/IAgentManager.cs
