+++ src/Agent/NewRelic/Agent/Core/RuntimeEnvironmentInfo.cs
