+++ src/Agent/NewRelic/Agent/Core/Singleton.cs
